Day 05 complete and changed Day template to be long part 1 and part 2

+++ src/main/template/Day.java
@@ -39,9 +39,9 @@ public abstract class Day {
         return input;
     }
 
-    public abstract int part1();
+    public abstract long part1();
 
-    public abstract int part2();
+    public abstract long part2();
 
 
 }
